requirements.txt
absl-py==2.1.0
attrs==23.1.0
biopython==1.83
Brotli==1.0.9
brotlipy==0.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.4
cloudpickle==2.2.1
comet-ml==3.34.0
configobj==5.0.8
cryptography==39.0.1
cycler==0.11.0
dulwich==0.21.6
einops==0.8.1
everett==3.1.0
fonttools==4.34.4
future==0.18.3
google==3.0.0
graphviz==0.8.4
greenlet==1.1.2
idna==3.4
importlib-resources==6.1.0
joblib==1.2.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
markdown-it-py==3.0.0
matplotlib==3.5.2
munkres==1.1.4
mxnet==1.9.1
networkx==3.1
numpy==1.24.4
packaging==23.1
pandas==1.4.3
Pillow==9.2.0
pip==23.3.2
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pooch==1.7.0
prettytable==2.2.1
psutil==5.9.6
py4j==0.10.9.7
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.8.2
pytz==2022.2
PyYAML==6.0.1
referencing==0.30.2
reportlab==3.5.68
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.6.0
rpds-py==0.10.6
scikit-learn==1.3.1
scipy==1.10.1
semantic-version==2.10.0
sentry-sdk==1.32.0
setuptools==68.0.0
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
SQLAlchemy==1.4.40
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
threadpoolctl==2.2.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
unicodedata2==14.0.0
urllib3==1.26.16
wcwidth==0.2.8
websocket-client==1.3.3
wheel==0.42.0
wrapt==1.15.0
wurlitzer==3.0.3
yacs==0.1.8
zipp==3.17.0
